requirements.txt
apache-airflow==2.8.1
pandas==2.1.4
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==1.4.53